First version of withOnyx HOC typing
