+++ lib/index.d.ts
@@ -1,7 +1,6 @@
 import Onyx from './Onyx';
+import {CustomTypeOptions} from './types';
 import withOnyx from './withOnyx';
-
-interface CustomTypeOptions {}
 
 export default Onyx;
 export {CustomTypeOptions, withOnyx};
